# contracts/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: GoldPC API
  description: |
    REST API для веб-приложения компьютерного магазина GoldPC с сервисным центром.
    
    ## Аутентификация
    API использует JWT-токены для аутентификации. Включите токен в заголовок:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Версионирование
    Текущая версия API: v1. Версионирование через URL путь (/api/v1/...).
    
    ## Формат дат
    Все даты и время в формате ISO 8601 (UTC).
    
    ## Коды ошибок
    - `400` - Некорректный запрос
    - `401` - Неавторизован
    - `403` - Доступ запрещён
    - `404` - Ресурс не найден
    - `409` - Конфликт
    - `422` - Ошибка валидации
    - `429` - Превышен лимит запросов
    - `500` - Внутренняя ошибка сервера
    
    ## Роли пользователей
    - **Client** — Клиент (оформление заказов, просмотр каталога)
    - **Manager** — Менеджер (обработка заказов, управление товарами)
    - **Master** — Мастер (выполнение услуг сервисного центра)
    - **Admin** — Администратор (полный доступ)
    - **Accountant** — Бухгалтер (финансовые отчёты)
  version: 1.0.0
  contact:
    name: GoldPC API Support
    email: api@goldpc.by
    url: https://goldpc.by/api-docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.goldpc.by/v1
    description: Production server
  - url: https://api-staging.goldpc.by/v1
    description: Staging server
  - url: http://localhost:5000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /wishlist/sync:
    put:
      tags: [Users]
      summary: Синхронизировать избранное текущего пользователя
      operationId: syncWishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
      responses:
        '200':
          description: Актуальный список избранного
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
  /wishlist/{productId}:
    post:
      tags: [Users]
      summary: Добавить товар в избранное
      operationId: addWishlistItem
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      responses:
        '200':
          description: Товар добавлен
    delete:
      tags: [Users]
      summary: Удалить товар из избранного
      operationId: removeWishlistItem
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      responses:
        '200':
          description: Товар удалён

  # === CATALOG ===
  /catalog/products:
    get:
      tags: [Catalog]
      summary: Получить список товаров
      description: Возвращает список товаров с пагинацией и фильтрацией
      operationId: getProducts
      security: []
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/Page'
        - $ref: './components/parameters/common.yaml#/components/parameters/PageSize'
        - $ref: './components/parameters/common.yaml#/components/parameters/Category'
        - $ref: './components/parameters/common.yaml#/components/parameters/Manufacturer'
        - $ref: './components/parameters/common.yaml#/components/parameters/PriceMin'
        - $ref: './components/parameters/common.yaml#/components/parameters/PriceMax'
        - $ref: './components/parameters/common.yaml#/components/parameters/Search'
        - $ref: './components/parameters/common.yaml#/components/parameters/SortBy'
        - $ref: './components/parameters/common.yaml#/components/parameters/SortOrder'
        - name: inStock
          in: query
          schema:
            type: boolean
          description: Только в наличии
        - name: isFeatured
          in: query
          schema:
            type: boolean
          description: Только рекомендуемые
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: './components/schemas/product.yaml#/components/schemas/ProductListResponse'
  /catalog/products/{productId}:
    get:
      tags: [Catalog]
      summary: Получить товар по ID
      operationId: getProduct
      security: []
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      responses:
        '200':
          description: Детали товара
          content:
            application/json:
              schema:
                $ref: './components/schemas/product.yaml#/components/schemas/Product'
        '404':
          $ref: './components/responses/common.yaml#/components/responses/NotFound'
  /catalog/products/{productId}/reviews:
    get:
      tags: [Catalog]
      summary: Получить отзывы о товаре
      operationId: getProductReviews
      security: []
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
        - $ref: './components/parameters/common.yaml#/components/parameters/Page'
        - $ref: './components/parameters/common.yaml#/components/parameters/PageSize'
      responses:
        '200':
          description: Список отзывов
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas/product.yaml#/components/schemas/Review'
                  meta:
                    $ref: './components/schemas/common.yaml#/components/schemas/PaginationMeta'

    post:
      tags: [Catalog]
      summary: Добавить отзыв
      operationId: addProductReview
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/product.yaml#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Отзыв добавлен
        '400':
          $ref: './components/responses/common.yaml#/components/responses/ValidationError'
        '401':
          $ref: './components/responses/common.yaml#/components/responses/Unauthorized'
...
  /catalog/manufacturers:
    get:
      tags: [Catalog]
      summary: Получить список производителей
      operationId: getManufacturers
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Фильтр по категории
      responses:
        '200':
          description: Список производителей
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas/product.yaml#/components/schemas/Manufacturer'

  # === PC BUILDER ===
...
  /admin/products:
    post:
      tags: [Admin]
      summary: Создать товар (Manager+)
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/product.yaml#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: './components/schemas/product.yaml#/components/schemas/Product'
        '400':
          $ref: './components/responses/common.yaml#/components/responses/ValidationError'
        '403':
          $ref: './components/responses/common.yaml#/components/responses/Forbidden'
  /admin/products/{productId}:
    put:
      tags: [Admin]
      summary: Обновить товар (Manager+)
      operationId: updateProduct
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/product.yaml#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Товар обновлён
          content:
            application/json:
              schema:
                $ref: './components/schemas/product.yaml#/components/schemas/Product'
        '400':
          $ref: './components/responses/common.yaml#/components/responses/ValidationError'
        '403':
          $ref: './components/responses/common.yaml#/components/responses/Forbidden'
        '404':
          $ref: './components/responses/common.yaml#/components/responses/NotFound'

    delete:
      tags: [Admin]
      summary: Удалить товар (Admin)
      operationId: deleteProduct
      parameters:
        - $ref: './components/parameters/common.yaml#/components/parameters/ProductId'
      responses:
        '204':
          description: Товар удалён
        '403':
          $ref: './components/responses/common.yaml#/components/responses/Forbidden'
        '404':
          $ref: './components/responses/common.yaml#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации